+++ packages/vue-instantsearch/src/components/HierarchicalMenu.vue
@@ -28,6 +28,7 @@
           !state.canToggleShowMore && suit('showMore', 'disabled'),
         ]"
         :disabled="!state.canToggleShowMore"
+        :aria-expanded="state.isShowingMore ? 'true' : 'false'"
         @click.prevent="state.toggleShowMore"
       >
         <slot name="showMoreLabel" :is-showing-more="state.isShowingMore">