+++ packages/algoliasearch-helper/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.29.2](https://github.com/algolia/instantsearch/compare/algoliasearch-helper@3.29.1...algoliasearch-helper@3.29.2) (2026-07-07)
+
+**Note:** Version bump only for package algoliasearch-helper
+
+
+
+
+
 # [3.29.0](https://github.com/algolia/instantsearch/compare/algoliasearch-helper@3.28.2...algoliasearch-helper@3.29.0) (2026-05-05)